贪心:统计当前字符串中符合pattern的个数，另外通过添加一个字符可以增多max(cnt_x,cnt_y)个 pattern

+++ src/leetcode/greedy/T2207.java
@@ -8,7 +8,7 @@ package leetcode.greedy;
 public class T2207 {
     public long maximumSubsequenceCount(String text, String pattern) {
         char x = pattern.charAt(0), y = pattern.charAt(1);
-        int cnt_x = 0, cnt_y = 0, ans = 0;
+        long cnt_x = 0, cnt_y = 0, ans = 0;
         for (int i = 0; i < text.length(); i++) {
             if (text.charAt(i) == y) {
                 ans += cnt_x;
